pipeline/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
pyarrow
altair
streamlit
plotly